Surgical gowns (#20555)
Adds surgical gowns, for use in operating rooms or sterile environments.
These come in 5 different corporate colors as well as a recolorable
generic variant.

Gowns appear on Surgeons by default, are available in the loadout (+
colorable variant), and in medical lockers alongside scrubs.

Deletes the surgical apron from the loadout. Still available in code but
supplanted entirely by the gowns now.

⚠️ **Gowns replace the labcoat on the Surgeon loadout** (to prevent
surgeons from forgetting them). This does unfortunately mean the labcoat
will need to be added in as a separate loadout item now to characters,
but the cost should be minimal.

Variants

![image](https://github.com/user-attachments/assets/7bd9c2bb-d787-4506-af0e-566abe2b4905)

Ingame

![image](https://github.com/user-attachments/assets/0fc46a0a-54b1-421a-a6ac-167071fc9038)

Colorable

![image](https://github.com/user-attachments/assets/c5c6b0fc-b76e-4113-839d-63deb972d01b)