Removes androids (not all cyborgs, just androids!) (#20577)
Per request of synthlore - you can no longer select the android job
title to spawn in with a positronic brain in a shipbound chassis, and
neither can you manufacture one.

Cyborgs broadly are not removed and are completely playable, but they
must contain either an MMI or a robot circuit. Positronics can still be
installed into an AI core.