Konyang planetside POIs alterations (#20587)
This alters some of the planetside POIs for Konyang. It removes the dead
IPCs from them, as well as makes it look like repair is underway on some
of the damaged buildings. It also removes the hivebots from two maps
that had those, since the military would likely have cleared those out
by now.