Ported radiation collector arrays from baystation12 (#19174)
These are intended to be available in the cargo list, so marking as
draft until that can be confirmed.

This ports radiation collectors from baystation12. These use the
ssradiation system that we have, to generate power from phoron and
radiation. This is contrasted with the prior iteration of these units,
which solely generated power from the singularity. These could be used
to generate power around a supermatter reactor, or a fusion reactor, or
any number of things that the playerbase can think of. The only thing is
that these cost precious phoron; you do not want to use these as the
sole source of power aboard anything.

### Asset Licenses
The following assets that **have not** been created by myself are
included in this PR:

| Path | Original Author | License |
| --- | --- | --- |
| icons/obj/machinery/rad_collector.dmi | Chinsky (Baystation12) |
CC-BY-SA |

---------

Co-authored-by: Cody Brittain <cbrittain10@live.com>