Antagland Spring Cleaning (#20601)
Does some spring cleaning on the off-ship antagonist area. This is
primarily in relation to lighting, decals, and some small changes to the
layouts of shuttles - particularly the mercenary shuttle, which has
received significant changes to its shape.

This is not a comprehensive remap, it's only to bring the area slightly
more in line with the mapping standards of the server.

I suspect that starlight doesn't render against unsimulated turfs, so
I've had to use some janky floating starlight fixtures outside of
windows to simulate that. It's not ideal, but neither is anything in CC.


![image](https://github.com/user-attachments/assets/d2bec39c-0e59-49eb-a1b2-76990e8f7170)


![image](https://github.com/user-attachments/assets/f7db446f-4bbb-490c-9cbe-b7f0e34606d0)


![image](https://github.com/user-attachments/assets/e196fd98-5840-4254-80e0-798ad135d49b)


![image](https://github.com/user-attachments/assets/0d910c23-5232-43e4-bdae-0f7443ae4402)


![image](https://github.com/user-attachments/assets/432538c5-2692-4636-aaa5-71b282550dfc)