Allows Chemistry Bags To Hold Inhaler Cartridges (#20606)
Given that the pharmacist is given inhalers and autoinhalers to fill
every round with pneumalin, pulmo, dexalin, and whatever, I figured this
might've been an oversight.

Super helpful for Odyssey rounds when all the crew are going down and
medical wants to bring the entire fridge (8 autoinhalers + 2 inhalers +
4-6 small inhaler cartridges AND the bottles)

EDIT: Changed to only inhaler cartridges, not autoinhalers nor inhaler
bases.