Adds Cigarette Cases to the loadout. (#20637)
Adds a new item to the smoking loadout tab, an empty metal cigarette
case that can hold 14 cigarettes. Intended to be filled up by rolling
people or those who prefer to keep all of their poison-in-a-stick within
a cute, convenient package.

Fixes an issue wherein the in-hand mob sprites would not display for
cigarette packets and cigar cases, they now do so properly. Modified the
north facing sprites so they don't look out of place (appear BEHIND the
player sprite)

Fixes an issue wherein you could place cigars into smaller cigarette
packets, which became relevant because that meant you could stuff the
cigarette case full of oversized cigars. Now you can only hold cigars in
cigar cases, as you should.

Built for modularity, it should be easy for contributors to add their
own cigarette cases design, and to tweak the existing cigar cases code
to allow for the same handling of open icon states, where multiple
closed icon sprites change to one single batch of open icon sprites,
reducing bloat. Courtesy of Geeves.

While currently out of scope of the P-R, this is intended to be improved
later down the line by allowing you to select which kinds of cigarettes
can spawn in the case in addition to allowing it to start empty, kind of
like how lunch boxes work. Was too complex for me to figure out and
would've taken far longer than needed, but is a planned project for the
future.