A small nerf to the "Creature" spawn (#20678)
This change reduces the damage that the green, watermelon-looking
"Creature" mobs are capable of.

Previously, their attacks had a damage range of 20-50. For comparison,
Plains Tyrant and Cavern Geist do 40 damage an attack.

The Xenoarchaology item that spawns these green dudes seems to be
capable of spawning them pretty quickly - and when you're fighting
simple mobs that are able to break your bones and give you an arterial
bleed in a single hit, that gets silly pretty fast.

I've tweaked the damage numbers so that the Creatures are a little
stronger than Hivebot Guardians in terms of damage.