Fixes Chainlink Fence Pathing (#20677)
Fixes an issue where you could not walk on the tile where a chainlink
fence existed. Altered the passing code and added some flags to ensure
it would check direction.

@RustingWithYou I tested it several times to make sure that it could not
be lept. It seemed the intention was for it to never be passable so
nothing living is able to move through it in this configuration. I left
the probability for projectiles to sometimes not pass.