Unapologetic Zeng-Hu vinyl gloves buff (it makes sense though) (#20687)
So Zeng-Hu, a medical and science company has gloves right? But they're
not on par with neither latex nor nitrile gloves, now that didn't make
sense to me so here's a buff, now they're on par with nitrile gloves.

(I don't know if glove powergaming happens but can reduce them to latex
if _nitrile_ is too strong)