Actor Preparation remap + clothes vendor (#20674)
Remaps the Actor's area of CC to be a little more inviting to players.
There aren't very many tangible changes to the balance, besides that the
areas separated from the rest by glass in the original map have been
removed, the contents of the medical section moved to an open area, and
the number of guns reduced and moved similarly to an open area.

Food and drinks are added as well, for players that want to pack their
lunches, and an incinerator is added so you can incinerate your trash.
Or other Actors, if you want to annoy the ST.

This also introduces a generic clothing vendor, populated primarily with
non faction-specific recolourable clothing items, bags, and accessories,
in addition to the clothes dyer item made for the Golden Deep ship so
you can actually recolour all those items. This is added to every
off-ship antag area.

Very few origin-specific items have been added because adding them would
increase the scope of the vendor so drastically that I feel we would
need a much better UI than the standard vendor UI currently being
utilised. It's not the best possible implementation of a way to give
antagonists access to generic clothes, but a clunky first iteration is
better than nothing.


![image](https://github.com/user-attachments/assets/98e77cc5-7b7c-4b37-9122-d578519f68c8)


![image](https://github.com/user-attachments/assets/1d487e77-39ad-42da-a9b0-e29b3d2b8845)