Hydrobuff v2 (#20634)
Adds a bunch of stuff to the bio-generator, remaps the garden and
hydroponics, adds some new magazines, modifies the plushies selection to
include more plushies, adds a few new bounties, and throws the kitchen a
bone by giving them soy sauce on round start.

Most controversial changes are likely to be:
- I removed back access to lower hydroponics for the cook/chef, and
added a windoor and table to discourage them accessing hydroponics.
- I made the garden smaller and made the vending machine now cost money
to get fertilizer and tools. The amount of credits should be fairly
reasonable, allowing most characters to still use the garden without
issue.

---------

Signed-off-by: courierbravo <73250577+courierbravo@users.noreply.github.com>
Co-authored-by: Wowzewow (Wezzy) <42310821+alsoandanswer@users.noreply.github.com>