Adds joy messages to Aphrodite (#20740)
Seeing as the drug has enough of an effect on people to cause factory
farming of diona and spawn a whole faction of diona birthed from such a
practice, it's probably safe to say it has some kind of pleasurable
effect when consumed.

This makes that a little more clear.