Arterial bleeding rebalance (#20680)
This reduces the multiplier for arterial bleeding blood a little, making
it 0.75 from 1 for the head and upper body, and a uniform 0.5 for all
limb organs, from 0.75 for arms and legs.

This is intended to keep arterials enough of a danger that they require
immediate medical treatment to resolve, while giving crew a little more
leeway with them before they go critical. This is mostly aimed at crew
without medical, including when there is no medical at all and also for
miners and other roles who leave the ship and need to come back to be
treated for arterials.

I have tested these new values, even 0.5 is enough to substantively
reduce blood volume over several minutes. The lower threshold for an AB
being inconsequential appears to be around 0.25, which is why I haven't
lowered any values that far down.