Add custom background highlights (#20701)
This is an update of
https://github.com/Aurorastation/Aurora.3/pull/19561 for BYOND 516.

Overall I think it's much cleaner. It makes use of CSS variables and
doesn't change the overall structure of the html like the last PR. Let
me know if any improvements can be made.


![image](https://github.com/user-attachments/assets/a5dc1dc2-3ca8-49f6-9f3e-01359f0cbf92)

Example. The left border and background highlight color will be the last
in the highlight settings.