Hydroponics layout tweaks (#20749)
This implements several tweaks to the layout and contents of the
recently remapped hydroponics bay to make the space smoother and more
intuitive to play in.

The intention of the mapping here is to make D1 hydroponics a
non-liminal space to be in, it should have amenities that mean you can
spend a long time there comfortably - I believe this was also the
intention of the original PR. It contains most of the hydroponics
machines. This is also intended to ensure that you can comfortably spend
most of your time in D2 if you prefer, which is important as it's the
more visible area.

The empty middle row in the D1 hydroponics racks is intended to be an
intuitive spot to cultivate bees. There are two chairs at the desk since
there are two gardeners, and the vendors have been moved to a more
prominent spot on D2 to make it feel less liminal. I've replaced the
hatches under the hydroponics trays with outlines, as a green hatch
under a plant makes the plant pretty indistinct. Windoors have been
removed between the two areas, since they are no longer required to gate
chef access.


![image](https://github.com/user-attachments/assets/40d393ec-4201-45b8-b710-9e155f36dfcc)

![image](https://github.com/user-attachments/assets/ca7b8c40-846d-4e10-bef0-1e3d5cfd897a)