Maintenance Panels openable with regular wrench (#20752)
To allow the fluff maintenance panels to be opened even with low-tech
standard wrenches. Not only does it make sense, but also lets them still
be opened and worked with even if Engineering were to lose its default
impact wrenches.

To facilitate the mundane wrench being used for atmos panels as a 'fix
RP' action, opening and closing maintenance panels now must be done with
the Disarm intent (with description updated to communicate the fact).

---------

Signed-off-by: Batrachophreno <Batrochophreno@gmail.com>
Co-authored-by: hazelrat <83198434+hazelrat@users.noreply.github.com>