Loadout drug changes (#20756)
This PR does the following changes:
It changes the name of the 'Smoking' tab in the loadout to 'Drugs and
Medicines'
It moves the psychiatric medicines from the Utility tab to this newly
renamed Drugs and Medicines tab
It adds a few basic OTC medicines, the same ones found in the OTC vendor
in the medical lobby (except for dexalin and dylovene)
It adds some legal recreational drugs that can be selected in the
loadout.

IMPORTANT: You will have to re-select items that you previously had in
the Smoking tab, and re-select your psychiatric medication.