Fixed the kitchen's first-aid closet having non-first aid items (#20776)
The first aid closet in the kitchen will no longer attempt to rub salt
into your wounds while you scramble around for bandages.

The game would think that the soy sauce bottles and plant were meant to
spawn inside the first-aid cabinet, leading to an amusing surprise
whenever you went for the first aid kit.

I moved the soy sauce bottles to the fridge, and deleted the potted
plant that sometimes spawns on that counter.