APC Item Interaction Cleanup (#20763)
Originally was trying to fix the issue where, if you first used a welder
on a broken APC like the Guide to Construction on the wiki says, it
instead kind of deletes the APC, making it very difficult to continue
dismantling/repairing it without alt-click shenanigans (and nearly
impossible if you also have the alt-click bug).

Think I have the bug resolved and also did some very minor
refactoring/commenting documentation to make the code easier to navigate
in the future.