Tea no longer poisons Dionae (#20743)
As in the title. I've been told this was implemented in the first place
because tea formerly cleared radiation, and without that mechanic as
context it doesn't make much sense. Will require some wiki edits.

I believe this is wanted by Diona lore, but I'll throw it into their
discord channel to verify and close it if they'd prefer that.

Resolves https://github.com/Aurorastation/Aurora.3/issues/16887