Produce Box Rebalance (#20783)
Removes the fruits from taking up valuable space in the produce boxes.
Pastries and cakes are pretty much a luxury item for the most part - and
it isn't too hard to grow them in the minor amounts that they're
actually needed in recipes.

Moves rice from the produce box to just normal-ass bags beside flour.
They're basically as important as flour at this point.