Adds Commonwealth of Hieroaetheria passports to the Diona loadout (#20780)
Added with the approval of GrandMasterChef!

Consortium of Hieroaetheria: "A passport issued to a citizen of the
Consortium of Hieroaetheria. A testament to inclusiveness anywhere its
owner takes with it."

![image](https://github.com/user-attachments/assets/62dad239-ef47-42f8-90a0-11c128076dd5)

Union of Gla'orr: "A passport issued to a citizen of the Union of
Gla'orr. Considered somewhat of an opposite of the Consortium passport
in terms of what travel it enables."

![image](https://github.com/user-attachments/assets/0b3c8adf-a678-447b-8140-f2c050e52901)

Eternal Republic of the Ekane: "A passport issued to a citizen of the
Eternal Republic of the Ekane, a mark of having committed oneself to the
Eternal wholly."

![image](https://github.com/user-attachments/assets/ac6cfa4f-d5c3-4bbf-83f1-fa49e4595e2c)

---------

Co-authored-by: Gabby <59633984+feartheblackout@users.noreply.github.com>