Split body scanner and cryopod sprites (#20794)
Port of https://github.com/Baystation12/Baystation12/pull/34190, though
a recolor and touchup was done by CormallenField/Steviii.

I've been concerned for a while that the cryopod and body scanner shared
the same sprites. For the most part, this merely bugged me, as generally
these machines wouldn't be in the same place, so the likelihood of
confusing the two is low. The merging of #20772 changed this, as there
is now a cryopod in medical.

This splits the cryopod, sleeper, and body scanner into their own files.
The body scanner now has a different sprite with a slightly different
shape, and a completely different coloration as well. I also ported the
animated (and smaller!) console from Baystation while I was at it.

Compared to the one in baystation, this has been recolored purple by
CormallenField/Steviii to fit better with our color theming for medical
equipment.

![image](https://github.com/user-attachments/assets/82faa75c-aca6-4ded-a6c3-4c9d9cc00c05)
Note: Since this image was taken, a fix was done to fix the missing bits
on the broken state.

I noticed some incorrect states for sleeper panel icons, and an unused
sleeper working animation. I fixed up and implemented these during the
transition.

### Asset Licenses
The following assets that **have not** been created by myself are
included in this PR:

| Path | Original Author | License |
| --- | --- | --- |
| icons/obj/machinery/bodyscanner.dmi | Remission (Baystation12) |
CC-BY-SA 3.0 |