Two new hairstyles (Punk Shaved, Sidecut Short Left/Right) (#20778)
Adds a short variant of the punk sidecut haircut, just trying out the
sprite editor to get a feel for it/make sure there's no extra tricky
workflow steps for very simple baseline sprites to start with. Unless
there's a place with a list of wanted/needed basic sprites to attack,
will try to think of useful things to make for more future practice.

![punk_shaved_sidecut_short](https://github.com/user-attachments/assets/bff49be0-52ed-4177-aab1-aa83910ab78a)