Adds Lyod Fatesayer Cards (#20782)
This PR adds a human lore item, "Fatesayer" cards(name may be subject to
change).
These are used by the people of the Lyod to tell your fate. 
They are a religious item for the loadout, intended for believers of the
shamanism interpretation of the Tribunal.

Could also maybe be added as a small addition to the wiki. The deck
description explains how to use them. The cards description explain what
they mean.

The cards in question:

![image](https://github.com/user-attachments/assets/1cbd5a82-8244-4d95-b706-78082ab5fc69)