(Re)Adding Soul: Title3 (#20811)
This PR (re?)adds Title3 as possible server music and to the jukebox
(not the audiconsoles).
This song is as old as the game. Personally I like it a lot and think we
can have it for the approx. 3 people who have the lobby music not muted.
I used an improved, remixed version of the classic Title3 (originally
"Tintin On The Moon", composed by Jeroen Tel), that is better quality
and not grating on your ears. Some of you may know this version from
SS14.
Tested and working locally.

All appropiate attributions have been made. It is all marked under
(YouTube's) CC BY (reuse allowed) license.

### Asset Licenses
The following assets that **have not** been created by myself are
included in this PR:

| Path | Original Author | License |
| sound/music/lobby/title3mk2 | Jeroen Tel (original)/PinkBlossom6
(remix) | CC0 (free reuse) |