Pharmacy Over The Counter Gaming - Mapping/Balance/Bugfix (#20834)
This PR was made at request of Kermit to slightly enhance the pharmacist
game-play loop by introducing selling medication OTC. Legally or
otherwise >;)

- Adds a quikpay to the pharmacy along with some shelving. Moves some
stuff around to accommodate this
- Fixes the hydroponics quikpay paying into the operations account
instead of the service account. That's a whoopsie on my part.
- Doubles the price of most drugs in the OTC/Wall vendor. Everything is
still very much affordable, but this should encourage pharmacist players
to undercut the vending machine without getting peanuts in return.

---------

Signed-off-by: courierbravo <73250577+courierbravo@users.noreply.github.com>
Co-authored-by: SleepyGemmy <99297919+SleepyGemmy@users.noreply.github.com>