Rename vars that override internal vars (#20837)
## About The Pull Request

OpenDream is adding `usr` and `args` to the `SoftReservedKeyword`
pragma: https://github.com/OpenDreamProject/OpenDream/pull/2307

This PR renames `var/usr` to `var/user` and `var/args` to
`var/arguments`.

## Why It's Good For The Game

This pragma is currently set as an error on Aurora, because using
internal var names for declared var names is cringe.