Raises the number of glasses and pitchers in the glasses vendor (#20865)
Recent changes moved all the glasses from the booze vendor to the
glasses vendor, but seem not to have brought over the number of pitchers
and regular glasses that used to be in there to compensate. This
reintroduces the lost glasses and pitchers so the bar doesn't run out so
easily.

---------

Signed-off-by: hazelrat <83198434+hazelrat@users.noreply.github.com>
Co-authored-by: SleepyGemmy <99297919+SleepyGemmy@users.noreply.github.com>