Appends the cans in a yoke to its name (#20877)
As title, up to two cans inside a yoke are automatically appended to its
name. This is capped at two so the name doesn't get really long. This
also updates if cans are removed from the yoke.

Currently there is no way to tell what is in a yoke without seeing its
icon, which is troublesome for bartenders who have stacks of the things.
This is intended to alleviate that issue.


![image](https://github.com/user-attachments/assets/31435e8c-5ffa-4ba4-9193-fd9cf0fa7f07)


![image](https://github.com/user-attachments/assets/3e63505c-8ce0-40de-8e53-b9040f1e72a5)