Bishop Screens! + Other Bishop related fixes (#20898)
Breaks the monopoly baselines had over changeable screens by letting
Bishops do it as well. Currently there are only two options asides from
the default, an exclamation mark and a cross, but adding more is fairly
simple.
Also fixes the default face being layered on top of the alternative
heads by restricting what screens are available to each head.
Removes the markings that related to the lights in the alternative heads
as this implementation renders them obsolete. As a bonus, Siseyo and
Sienna lights are now animated!

The new screen options:

![image](https://github.com/user-attachments/assets/68c4a2eb-13f5-4ddb-8afb-9953e599b89d)