Generic Augment Customization (#20859)
This PR makes small adjustments to the code for the cyborg analyzer
tool, making it so that it also provides the description given for
augments taken as part of a character's loadout. Additionally, this sets
loadout flags for most generic augments (anything not explicitly tied to
a specific corporation) to allow for custom names and descriptions.

Lastly, I've extended the analyzer in its human-mode to also display
assisted implants. While I can't imagine a machinist would have much to
do with a pacemaker, it is a metal implant. I actually mainly had in
mind that I was annoyed with the positronic brain implant not showing up
on cyborg analyzers, and wanted to correct the injustice.