Remaps Intrepid and ports CM's double seats (#20809)
## About PR

- Introduces a dropship-esque Intrepid, making it feel more like a
public transport in a overpopulous city, while trying to avoid
chokepoints.

- Ports the CM's double seats.

- Removes access restriction from Intrepid's side airlocks, to be in par
with the main airlock.

- Deletes some rogue pipes near hangar bay.

- Re-adds a small oven which was introduced in #16449. With slight
tweaks and different colour palette.

This is probably an awful way to port CM's vehicle seats, but in testing
only issue I've found is: when the chair occupant punches someone, their
own buckle off-set will get reset. I am not sure why.

The way the double seats work is, it'll let anyone pass through as long
as both of the seats aren't occupied at the same time.

## Images

<details><summary>Details</summary>
<p>


![a](https://github.com/user-attachments/assets/19813e63-a773-46a7-9f2c-43b5a23906ab)


</p>
</details>

---------

Signed-off-by: Kano <89972582+kano-dot@users.noreply.github.com>
Co-authored-by: hazelrat <83198434+hazelrat@users.noreply.github.com>