The Shortening of the Way: Cryo Outpost & Nuclear Silo (#20805)
This partially remaps the Cryo Outpost and Nuclear Silo scenario maps to
reduce the distance between the shuttle landmarks and wherever anything
is happening.

The massive distances in the current versions isolate the crew from the
gimmick and frustrate their ability to get involved, particularly
impacting departments like operations and service which may want to stay
near the shuttle. This is intended to help remedy that.

Ideally I'd prefer the landmarks for both of these to be even closer
than these remaps place them, but this is the closest possible without
substantially altering elements of the map, such as the flashy submap in
the middle of the cryo outpost.