Diona magboots 2.0 (#20744)
This is a continuation of
https://github.com/Aurorastation/Aurora.3/pull/18875, full credit goes
to Matt for the original PR.

This adds an ability allowing Dionae to access the functionality of
magboots. As the only spaceborne species in the setting, they should
have more abilities allowing them to excel in their natural environment.

A slowdown has been added to using the ability that wasn't in the
original PR to establish a downside to using it, hopefully addressing
the feedback to the original PR. It can also now be toggled in space,
since I found it pretty frustrating having to find a floor turf every
time I was scrambling about on lattices wanting to toggle my roots.