Implement forum-user-auth for game (#20875)
Requires https://github.com/Aurorastation/Web-Interface/pull/136 be
merged.

Allows users to authenticate themselves using only the forum. It will
generate a custom ckey for these users in the form of
`GuestF-[ckey(forum-nickname)]`. This will be saved into the database
and will act as a normal ckey from thereon out. These ckeys will be
marked in the player table, to allow for later filtering.

This will also modify the config parameter external_auth to be an enum
of 3 values.

Open questions: can we do automatic linking from purely the game side?
Since upon creation of the custom ckey, it should also be linked back to
the forums.

---------

Co-authored-by: Erki <erki@skullnet.me>